pyproject.toml
[tool.poetry]
name = "pandasai-chainlit-docker-template"
version = "0.1.0"
description = "A template PandasAI Chainlit App with Docker Deployments"
authors = ["Amjad Raza"]
license = "MIT"
readme = "README.md"
packages = [{include = "pandasai_chainlit_docker_deployment_template"}]


[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
chainlit = "^1.1.202"
pyyaml = "^6.0.1"
pandasai = {extras = ["bedrock", "connectors"], version = "^2.0.43"}
boto3 = "^1.34.111"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
